fix index types

+++ Visualizer/AIPlayer.cpp
@@ -34,8 +34,8 @@ void AIPlayer::findMove(const Desk& d, int& xOpt, int& yOpt) {
 float AIPlayer::evaluateBoard(const Desk& d) const {
     Matrix input(65, 1);
 
-    for (int x = 0; x < 8; ++x) {
-        for (int y = 0; y < 8; ++y) {
+    for (size_t x = 0; x < 8; ++x) {
+        for (size_t y = 0; y < 8; ++y) {
             input(8 * x + y, 0) = float(d(x, y));
         }
     }
